Refactored helper components into their own files
