+++ app/assets/javascripts/Components/Assessment_Chart/core_statistics.jsx
@@ -1,0 +1,52 @@
+import React from "react";
+import {FractionStat} from "./fraction_statistic";
+import PropTypes from "prop-types";
+
+export class CoreStatistics extends React.Component {
+  render() {
+    const max_mark_value = Number(this.props.max_mark) || 0;
+    let percent_standard_deviation = "0.00";
+    if (max_mark_value !== 0) {
+      percent_standard_deviation = (
+        (100 / max_mark_value) *
+        (Number(this.props.standard_deviation) || 0)
+      ).toFixed(2);
+    }
+    let num_fails = "";
+    if (this.props.num_fails !== undefined) {
+      num_fails = (
+        <React.Fragment>
+          <span className="summary-stats-label">{I18n.t("num_failed")}</span>
+          <FractionStat numerator={this.props.num_fails} denominator={this.props.num_groupings} />
+        </React.Fragment>
+      );
+    }
+
+    return (
+      <React.Fragment>
+        <span className="summary-stats-label">{I18n.t("average")}</span>
+        <FractionStat numerator={this.props.average} denominator={this.props.max_mark} />
+        <span className="summary-stats-label">{I18n.t("median")}</span>
+        <FractionStat numerator={this.props.median} denominator={this.props.max_mark} />
+        <span className="summary-stats-label">{I18n.t("standard_deviation")}</span>
+        <span>
+          {(this.props.standard_deviation || 0).toFixed(2)}
+          &nbsp;({percent_standard_deviation}%)
+        </span>
+        {num_fails}
+        <span className="summary-stats-label">{I18n.t("num_zeros")}</span>
+        <FractionStat numerator={this.props.num_zeros} denominator={this.props.num_groupings} />
+      </React.Fragment>
+    );
+  }
+}
+
+CoreStatistics.propTypes = {
+  average: PropTypes.number.isRequired,
+  median: PropTypes.number.isRequired,
+  standard_deviation: PropTypes.number.isRequired,
+  max_mark: PropTypes.number.isRequired,
+  num_fails: PropTypes.number,
+  num_zeros: PropTypes.number.isRequired,
+  num_groupings: PropTypes.number.isRequired,
+};
