+++ app/assets/javascripts/Components/Assessment_Chart/fraction_statistic.jsx
@@ -1,0 +1,19 @@
+import React from "react";
+
+export class FractionStat extends React.Component {
+  render() {
+    const numerator = +(Number(this.props.numerator) || 0).toFixed(2),
+      denominator = +(Number(this.props.denominator) || 0).toFixed(2);
+    let result = "0.00";
+    if (denominator !== 0) {
+      result = ((Number(this.props.numerator) / Number(this.props.denominator) || 0) * 100).toFixed(
+        2
+      );
+    }
+    return (
+      <span>
+        {numerator} / {denominator} ({result}%)
+      </span>
+    );
+  }
+}
