+++ app/assets/stylesheets/common/_statistics.scss
@@ -43,8 +43,8 @@
     font-size: 1vw;
   }
 
-  .criteria-summary-table {
-    .criteria-stat-breakdown {
+  .grade-breakdown-summary-table {
+    .grade-stats-breakdown {
       column-gap: 10px;
       margin-bottom: 10px;
       margin-left: 40px;
